conectado a MongoDbAtlas

+++ src/seed/seed.service.ts
@@ -1,13 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Pokemon } from 'src/pokemon/entities/pokemon.entity';
 import { PokeResponse } from './interfaces/poke.response.interface';
 import { HttpAdapter } from 'src/common/adapters/http.adapter';
+import { ConfigService } from '@nestjs/config';
 
 
 @Injectable()
 export class SeedService {
+
+  private environment: string;
 
   constructor(
 
@@ -15,9 +18,15 @@ export class SeedService {
 
     @InjectModel(Pokemon.name)
     private readonly pokemonModel: Model<Pokemon>,
-  ) { }
+
+    private readonly configService: ConfigService
+  ) {
+    this.environment = configService.get<string>('environment') ?? '';
+  }
 
   async executeSeed() {
+
+    if (this.environment.toLocaleLowerCase() !== 'dev') throw new BadRequestException('No es posible aplicar el seed en producción')
 
     await this.pokemonModel.deleteMany({});
